+++ CHANGELOG.rst
@@ -10,6 +10,8 @@ releases, in reverse chronological order.
 * Some factories were not fully reusable since they depended on data files that
   were no included in this package. These files are now included, and all
   factories should now be reusable by downstream apps.
+* Fixed a bug when validating credit notes and other receipts which have a
+  related-receipt.
 
 8.0.2
 -----